controller: adjust mode and window selection
to better handle current conditions. Switching the mode should happen
before it gets too hot or too cold inside. Its currently very hot out (heat wave), but the system is still in heating mode because the indoor temp
has not yet gone over the cool setpoint. Will now use the mid point between heat and cool setpoints.

Because the system was still in heating mode, it recommended to open the windows! It was too hot and too humid out for that to be a good idea.